requirements-test.txt
pytest-django
factory-boy
mock
selenium
